// galaxyui/src/app/resources/provider-namespaces/provider-source.ts
export class ProviderSource {
    summary_fields: any;
    related: any;
    name: string;
    description: string;
    display_name: string;
    avatar_url: string;
    location: string;
    company: string;
    email: string;
    html_url: string;
    followers: number;
    provider: number;
    provider_name: string;
    selected: boolean;
}